chore: moves the public loop mixin into its own file
