+++ core/blockly.js
@@ -43,6 +43,7 @@ const dialog = goog.require('Blockly.dialog');
 const fieldRegistry = goog.require('Blockly.fieldRegistry');
 const geras = goog.require('Blockly.geras');
 const internalConstants = goog.require('Blockly.internalConstants');
+const loopMixin = goog.require('Blockly.loopMixin');
 const minimalist = goog.require('Blockly.minimalist');
 const registry = goog.require('Blockly.registry');
 const svgMath = goog.require('Blockly.utils.svgMath');
@@ -682,6 +683,7 @@ exports.fieldRegistry = fieldRegistry;
 exports.geras = geras;
 exports.inject = inject;
 exports.inputTypes = inputTypes;
+exports.loopMixin = loopMixin;
 exports.minimalist = minimalist;
 exports.registry = registry;
 exports.thrasos = thrasos;
