+++ externs/generator-externs.js
@@ -12,6 +12,7 @@
 goog.provide('Blockly');
 goog.provide('Blockly.Generator');
 goog.provide('Blockly.inputTypes');
+goog.provide('Blockly.loopMixin');
 goog.provide('Blockly.utils.global');
 goog.provide('Blockly.utils.object');
 goog.provide('Blockly.utils.string');
