+++ generators/dart/loops.js
@@ -150,7 +150,7 @@ Blockly.Dart['controls_flow_statements'] = function(block) {
     xfix += Blockly.Dart.injectId(Blockly.Dart.STATEMENT_SUFFIX, block);
   }
   if (Blockly.Dart.STATEMENT_PREFIX) {
-    const loop = Blockly.Constants.Loops
+    const loop = Blockly.loopMixin
         .CONTROL_FLOW_IN_LOOP_CHECK_MIXIN.getSurroundLoop(block);
     if (loop && !loop.suppressPrefixSuffix) {
       // Inject loop's statement prefix here since the regular one at the end
