+++ generators/lua/loops.js
@@ -154,7 +154,7 @@ Blockly.Lua['controls_flow_statements'] = function(block) {
     xfix += Blockly.Lua.injectId(Blockly.Lua.STATEMENT_SUFFIX, block);
   }
   if (Blockly.Lua.STATEMENT_PREFIX) {
-    const loop = Blockly.Constants.Loops
+    const loop = Blockly.loopMixin
         .CONTROL_FLOW_IN_LOOP_CHECK_MIXIN.getSurroundLoop(block);
     if (loop && !loop.suppressPrefixSuffix) {
       // Inject loop's statement prefix here since the regular one at the end
